コンポーネントの分割
